// src/models/enums/MessageEnumeration.ts
enum MessageEnumeration {
    REGISTER_CREATED = "Registro criado com sucesso.",
    REGISTER_NOT_FOUND = "Registro não encontrado.",
    USER_NOT_EXIST = "Usuário não existe.",
    COURSE_NOT_EXIST = "Curso não existe.",
    ENROLLMENT_NOT_EXIST = "Matrícula não existe.",
    REGISTER_ALREADY_EXIST = "Registro já existe.",
    NOT_AUTHORIZED = "Acesso não autorizado.",
}

export default MessageEnumeration;